[Rust][doc](global, lib): Added doc comments for Global and GlobalType;
