+++ bindings/rust/wasmedge-sys/src/instance/mod.rs
@@ -1,9 +1,15 @@
+//! Defines WasmEdge instance structs, including Function, Global, Memory, and Table.
+
 pub mod function;
 pub mod global;
 pub mod memory;
 pub mod table;
 
-pub use function::Function;
+#[doc(inline)]
+pub use function::{FuncType, Function};
+#[doc(inline)]
 pub use global::{Global, GlobalType};
-pub use memory::Memory;
-pub use table::Table;
+#[doc(inline)]
+pub use memory::{MemType, Memory};
+#[doc(inline)]
+pub use table::{Table, TableType};
